+++ tests/smoke/test_imports.py
@@ -91,9 +91,9 @@ def test_rdbms_data_model_imports():
 
 
 def test_enrichment_imports():
-    from neocarta.enrichment.embeddings import OpenAIEmbeddingsConnector
+    from neocarta.enrichment.embeddings import LiteLLMEmbeddingsConnector
 
-    assert OpenAIEmbeddingsConnector
+    assert LiteLLMEmbeddingsConnector
 
 
 def test_ingest_imports():